+++ otserv/trunk/spells.cpp
@@ -245,7 +245,7 @@ SpellScript::SpellScript(const std::string &datadir, std::string scriptname, Spe
 	lua_dofile(luaState, scriptname.c_str());
 	this->loaded=true;
 	this->spell=spell;
-	this->setGlobalNumber("addressOfSpell", (int) spell);
+	this->setGlobalNumber("addressOfSpell", (long) spell);
 	this->registerFunctions();
 }
